+++ resources/js/requests/components/RequestsListing.vue
@@ -142,6 +142,7 @@ export default {
   },
   mounted() {
     this.setupColumns();
+    this.getParticipants("");
   },
   methods: {
     setupColumns() {
