+++ resources/js/tasks/components/PMColumnFilterPopoverTasksMixin.js
@@ -1,35 +1,15 @@
+import PMColumnFilterPopoverCommonMixin from "../../common/PMColumnFilterPopoverCommonMixin.js";
+
 const PMColumnFilterPopoverTasksMixin = {
+  mixins: [PMColumnFilterPopoverCommonMixin],
   data() {
     return {
-      advancedFilter: [],
       viewAssignee: []
     };
   },
   methods: {
     onChangeSort(value) {
       this.order_direction = value;
-    },
-    onApply(json, index) {
-      this.advancedFilter[index] = json;
-      this.tableHeaders[index].filterApplied = true;
-      this.fetch();
-    },
-    onClear(index) {
-      this.advancedFilter[index] = [];
-      this.tableHeaders[index].filterApplied = false;
-      this.fetch();
-    },
-    onUpdate(object, index) {
-      if (object.$refs.pmColumnFilterForm &&
-              this.advancedFilter.length > 0 &&
-              this.advancedFilter[index] &&
-              this.advancedFilter[index].length > 0) {
-        object.$refs.pmColumnFilterForm.setValues(this.advancedFilter[index]);
-      }
-    },
-    getAdvancedFilter() {
-      let flat = this.advancedFilter.flat(1);
-      return flat.length > 0 ? "&advanced_filter=" + JSON.stringify(flat) : "";
     },
     getFormat(column) {
       let format = "string";
@@ -47,7 +27,7 @@ const PMColumnFilterPopoverTasksMixin = {
         formatRange = this.getStatus();
       }
       if (column.field === "assignee") {
-        formatRange = this.getAssignee();
+        formatRange = this.viewAssignee;
       }
       return formatRange;
     },
@@ -58,91 +38,14 @@ const PMColumnFilterPopoverTasksMixin = {
       }
       return operators;
     },
-    getViewConfigFilter() {
-      return [
-        {
-          "type": "string",
-          "includes": ["=", "<", "<=", ">", ">=", "contains", "regex"],
-          "control": "PMColumnFilterOpInput",
-          "input": ""
-        },
-        {
-          "type": "string",
-          "includes": ["between"],
-          "control": "PMColumnFilterOpBetween",
-          "input": []
-        },
-        {
-          "type": "string",
-          "includes": ["in"],
-          "control": "PMColumnFilterOpIn",
-          "input": []
-        },
-        {
-          "type": "datetime",
-          "includes": ["=", "<", "<=", ">", ">=", "contains", "regex"],
-          "control": "PMColumnFilterOpDatetime",
-          "input": ""
-        },
-        {
-          "type": "datetime",
-          "includes": ["between"],
-          "control": "PMColumnFilterOpBetweenDatepicker",
-          "input": []
-        },
-        {
-          "type": "datetime",
-          "includes": ["in"],
-          "control": "PMColumnFilterOpInDatepicker",
-          "input": []
-        },
-        {
-          "type": "stringSelect",
-          "includes": ["="],
-          "control": "PMColumnFilterOpSelect",
-          "input": ""
-        },
-        {
-          "type": "stringSelect",
-          "includes": ["in"],
-          "control": "PMColumnFilterOpSelectMultiple",
-          "input": []
-        },
-        {
-          "type": "boolean",
-          "includes": ["="],
-          "control": "PMColumnFilterOpBoolean",
-          "input": false
-        }
-      ];
-    },
-    getStatus() {
-      //todo: add labels "In Progress", "Completed", "Error", "Canceled"
-      return ["ACTIVE", "FAILING", "COMPLETED", "CLOSED", "EVENT_CATCH", "TRIGGERED", "INCOMING"];
-    },
-    getAssignee() {
+    getAssignee(filter) {
       //todo: Users can be numerous; therefore, the control should be a search 
       //field with a suggestion list type.
-      if (this.viewAssignee.length > 0) {
-        return this.viewAssignee;
-      }
-      let page = 1;
-      let perPage = 100;
-      let filter = "";
-      let orderBy = "username";
-      let orderDirection = "asc";
-      let url = "users" +
-              "?page=" + page +
-              "&per_page=" + perPage +
-              "&filter=" + filter +
-              "&order_by=" + orderBy +
-              "&order_direction=" + orderDirection;
-      ProcessMaker.apiClient.get(url).then(response => {
+      ProcessMaker.apiClient.get(this.getUrlUsers(filter)).then(response => {
         for (let i in response.data.data) {
           this.viewAssignee.push(response.data.data[i].username);
         }
       });
-      return this.viewAssignee;
     }
   }
 };
